% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_dataframe}
\alias{as_dataframe}
\title{Convert the projections to a data.frame}
\usage{
as_dataframe(x)
}
\arguments{
\item{x}{output from \code{spatial_project}}
}
\value{
data.frame with columns
sim, day, and 1 column with each location.
}
\description{
This function convers the output of
\code{spatial_project} function into a wide
data.frame.
}
\author{
Sangeeta Bhatia, Anne Cori, Pierre Nouvellet
}
