% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_of_infection.R
\name{force_of_infection}
\alias{force_of_infection}
\title{Force of infection under a spatially explicit brancing process model}
\usage{
force_of_infection(x, r_t, ws, pmovement)
}
\arguments{
\item{x}{A matrix of past incidence (integer). The matrix has 1
column for each location; that is, each column is interpreted as
the incidence in a location.}

\item{r_t}{A matrix of reproduction numbers at times 1 through t
in locations 1 through n.}

\item{ws}{flipped serial interval}

\item{pmovement}{A n_locations X n_locations matrix of the
probability of movement between locations. Entry in row i, column j
is the probability that a case in i will move to j during their
infectious period.}
}
\value{

}
\description{
This function computes the force of infection at time t + 1 at
locations 1, 2,....n
given incidence (by onset) at these locations, the serial interval
and the probabilities of movement of individuals during their
infectious period.
}
\author{
Sangeeta Bhatia, Anne Cori, Pierre Nouvellet
}
