% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Centroids and populations of all countries}
\format{
A dataframe consisting of 55 rows and 23 columns. The
columns of interest are Centroid_Lon, Centroid_Lat, and Pop
}
\source{
LandScan 2015 dataset grid
}
\description{
This data set gives the latitudes and longitudes for the
centroids of all countries on the African mainland. Also included are population
sizes. These data are needed to jointly estimate
reproduction numbers and parameters of movement
model in the spatially explicit brancing process
model implemented in this package.
}
\references{
Bhatia, S., Lassmann, B., Cohn, E. et al. Using digital surveillance tools for near real-time mapping of the risk of infectious disease spread. npj Digit. Med. 4, 73 (2021). https://doi.org/10.1038/s41746-021-00442-3
}
