% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promed_ebola}
\alias{promed_ebola}
\title{Processed data on the 2013–2016 West African Ebola epidemic}
\format{
A dataframe consisting of 656 rows and 56 columns. The
first column consists of dates and the remaining 55 columns contain
the incidence of Ebola cases for each country on the African mainland.
Column names are the ISO3 country codes of countries.
}
\source{
Bhatia, S., Lassmann, B., Cohn, E. et al. Using digital surveillance tools for near real-time mapping of the risk of infectious disease spread. npj Digit. Med. 4, 73 (2021). https://doi.org/10.1038/s41746-021-00442-3
}
\description{
This data set gives the daily incidence of
Ebola cases (includes suspected and confirmed cases)
for each country in Africa. Note that most of the
countries did not experience any Ebola cases in
this epidemic. The '0s' have been filled in
for convenience. The raw data were colleced by
ProMED and the pre-processing pipeline has been
described in detail elsewhere (See linked paper).
}
\references{
Bhatia, S., Lassmann, B., Cohn, E. et al. Using digital surveillance tools for near real-time mapping of the risk of infectious disease spread. npj Digit. Med. 4, 73 (2021). https://doi.org/10.1038/s41746-021-00442-3
}
