% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sd2}
\alias{sd2}
\title{Standard deviation with optional args}
\usage{
sd2(vec, ...)
}
\arguments{
\item{vec}{a numeric vector}

\item{...}{optional arguments}
}
\value{
numeric. standard deviation
}
\description{
This function computes the standard deviation of the values in x.
It calls \code{stats::sd} except that it also accepts optional
arguments and therefore can be used in conjunction with other functions
in \code{summarise_projections}.
}
\author{
Sangeeta Bhatia, Anne Cori, Pierre Nouvellet
}
