% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{spatial_estimater}
\alias{spatial_estimater}
\title{Estimate effective reproduction number using a spatially explicit
branching process model.}
\usage{
spatial_estimater(
  x,
  si,
  pmovement = NULL,
  tstart,
  tend,
  model = "poisson",
  priors,
  ...
)
}
\arguments{
\item{x}{A matrix of past incidence (integer).
The matrix has 1 column for each location; that is, each column is
interpreted as the incidence in a location.}

\item{si}{A matrix of discretised serial intervals. They will
usually be the same for each location.}

\item{pmovement}{A n_locations X n_locations matrix of the
probability of movement between locations. Entry in row i, column j
is the probability that a case in i will move to j during their
infectious period.}

\item{tstart}{A matrix of positive integers giving the starting
index of the windows over which reproduction number is estimated.}

\item{tend}{A matrix of positive integers giving the indices of the
endpoints of the windows over which reproduction number is estimated.}

\item{model}{Currently only "poisson" is supported}
}
\value{

}
\description{
Given past incidence in N locations, this function (a) jointly
estimates the reproduction number and the probability of movement,
or (b) if pmovement is not NULL, estimates only the reproduction
number
}
\author{
Sangeeta Bhatia
}
