% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{spatial_priors}
\alias{spatial_priors}
\title{Default priors for parameters of a spatially explicit branching process
model}
\usage{
spatial_priors()
}
\value{
a list of default parameters for the priors.
Values can then be manually be edited as in the examples below.
}
\description{
The parameters are the effective reproduction number and the exponent on the
distance in a gravity model, when both are being jointly estimated.
Effective reproduction number is assumed to be Gamma distributed
and therefore, its mean and standard deviation need to be specified.
These are automataically translated into the shape and scale paramaters.
When estimating only the effective reproduction number, the
`gamma` component of the list is silently ignored.
}
\examples{
priors <- spatial_priors()
priors$gamma <- 2
}
\author{
Sangeeta Bhatia
}
