% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialbranchr-package.R
\docType{package}
\name{spatialbranchr-package}
\alias{spatialbranchr-package}
\alias{spatialbranchr}
\title{The 'spatialbranchr' package.}
\description{
spatialbranchr implements a spatially explicit branching process model
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
