% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_projections}
\alias{summarise_projections}
\title{Summarise projections across simulations}
\usage{
summarise_projections(
  x,
  funs = c(quantile = quantile, min = min, max = max, mean = mean, sd = sd2),
  ...
)
}
\arguments{
\item{x}{output from \code{spatial_project} function}

\item{funs}{a list of summary functions to be applied.
It is the user's responsibility to ensure that the functions
are valid and can be applied to a vector. Defaults to
quantile, min, max, mean and sd.}

\item{...}{Additional arguments for funs, e.g. na.rm, probs}
}
\value{
A list of the same length as that of the supplied functions.
Each element of the list is the result of applying the corresponding function
across simulations for each day and each location.
}
\description{
Extract summary statistics for projections
}
\details{
If you want to use multiple functions and  pass optional
arguments other than na.rm, then use \code{sd2} rather than \code{sd}
to estimate standard deviation. This is because \code{sd} does not
accept optional arguments.
}
\author{
Sangeeta Bhatia, Anne Cori, Pierre Nouvellet
}
