% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{make_rindex}
\alias{make_rindex}
\title{Prepare index matrix for Rt}
\usage{
make_rindex(rindex, ncol, nrow, change_at)
}
\arguments{
\item{rindex}{vector of integer indices}

\item{ncol}{number of columns. Should be
the same as the number of locations.}

\item{nrow}{number of rows. Should be the same
as the number of time steps.}

\item{change_at}{A vector of integers indicating
for each location, at which time step should R
index change. Values are recycled.}
}
\value{
matrix of R index
}
\description{
Prepare index matrix for Rt
}
\details{
This is a utility function that prepared the
index matrix for Rt estimation in Stan.
Say we have two locations and a time series
of cases over 14 days. For each location,
we want to estimate Rt over a 7 day window
so that in all the model has 4 parameters for
Rt (two for each location). The matrix looks
like:
 \tabular{rr}{
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4
}
Given a vector R return a nrow * ncol matrix where
R[1:ncol] will be inserted in rows 1 to change_at[1].
Similarly R[ncol + 1: (2 * ncol)] will be inserted from
change_at[1] + 1 to row_indices[2] and so on.
}
\examples{
make_rindex(c(1, 2, 3), 3, 4, c(4))
make_rindex(c(1, 2, 3), 1, 9, c(5, 7))
}
\author{
Sangeeta Bhatia, Anne Cori, Pierre Nouvellet
}
