% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{make_rindex}
\alias{make_rindex}
\title{Prepare index matrix for Rt}
\usage{
make_rindex(window, nloc, T)
}
\arguments{
\item{window}{integer indicating the length of the window over
which Rt is to be estimated. Window length is assumed to be the same
for all locations. Default value is 7 i.e. Rt is estimated over
a weekly window. Currently only non-overlapping windows
are supported.}

\item{nloc}{integer. Number of locations.}

\item{T}{integer. Number of time steps}
}
\value{
matrix of R index
}
\description{
Prepare index matrix for Rt
}
\details{
This is a utility function that prepares the
index matrix for Rt estimation in Stan.
Say we have two locations and a time series
of cases over 14 days. For each location,
we want to estimate Rt over a 7 day window
so that in all, the model has 4 parameters for
Rt (two for each location). The matrix looks
like:
 \tabular{rr}{
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  1 \tab 3\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4\cr
  2 \tab 4
}
The columns index the locations and the rows index the time points.
}
\examples{
make_rindex(7, 2, 14)
make_rindex(4, 2, 14)
}
\author{
Sangeeta Bhatia, Anne Cori, Pierre Nouvellet
}
